API request prepared

+++ app/src/main/java/com/mobven/fitai/data/model/dto/FirstLoginDto.kt
@@ -2,9 +2,9 @@ package com.mobven.fitai.data.model.dto
 
 data class FirstLoginDto(
     val dateOfBirth: String,
-    val firstWeight: Int,
+    val firstWeight: Double,
     val gender: String,
     val goals: String,
-    val height: Int,
-    val targetWeight: Int
+    val height: Double,
+    val targetWeight: Double
 )
